source/requirements.txt
wheel==0.41.2
pip==23.2.1
Pillow==9.5.0
PyQt5==5.15.7
pygit2==1.12.1